(PC-15195)[API] refactor: move phone_validation function in subscription module
