+++ api/src/pcapi/admin/validators.py
@@ -1,6 +1,6 @@
 from wtforms.validators import ValidationError
 
-from pcapi.core.users.exceptions import InvalidPhoneNumber
+from pcapi.core.subscription.phone_validation import exceptions as phone_validation_exceptions
 from pcapi.utils import phone_number
 
 
@@ -10,5 +10,5 @@ class PhoneNumberValidator:
         if value:
             try:
                 phone_number.parse_phone_number(value, "FR")
-            except InvalidPhoneNumber:
+            except phone_validation_exceptions.InvalidPhoneNumber:
                 raise ValidationError("Numéro de téléphone invalide")
