+++ api/src/pcapi/core/subscription/phone_validation/constants.py
@@ -1,0 +1,14 @@
+WHITELISTED_COUNTRY_PHONE_CODES = {
+    33,  # France métropolitaine
+    590,  # Guadeloupe
+    596,  # Martinique
+    594,  # Guyane
+    262,  # Réunion
+    508,  # Saint-Pierre-et-Miquelon
+    262,  # Mayotte
+    590,  # Saint-Barthélémy
+    590,  # Saint-Martin
+    681,  # Wallis-et-Futuna
+    689,  # Tahiti
+    687,  # Nouvelle-Calédonie
+}
