+++ api/src/pcapi/core/users/constants.py
@@ -15,20 +15,6 @@ ACCOUNT_CREATION_MINIMUM_AGE = 15
 
 ACCOUNT_REACTIVATION_LIMIT = 60  # days
 
-WHITELISTED_COUNTRY_PHONE_CODES = {
-    33,  # France métropolitaine
-    590,  # Guadeloupe
-    596,  # Martinique
-    594,  # Guyane
-    262,  # Réunion
-    508,  # Saint-Pierre-et-Miquelon
-    262,  # Mayotte
-    590,  # Saint-Barthélémy
-    590,  # Saint-Martin
-    681,  # Wallis-et-Futuna
-    689,  # Tahiti
-    687,  # Nouvelle-Calédonie
-}
 
 EDUCONNECT_SAML_REQUEST_ID_TTL = 24 * 60 * 60  # 1 day in seconds
